// frontend/src/api/generated/models/gameResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { GameResponseDtoStatus } from "./gameResponseDtoStatus";
import type { GameTeamResponseDto } from "./gameTeamResponseDto";
import type { CategoryResponseDto } from "./categoryResponseDto";
import type { GameCategoryBoardResponseDto } from "./gameCategoryBoardResponseDto";

export interface GameResponseDto {
  _id: string;
  name: string;
  status: GameResponseDtoStatus;
  teams: GameTeamResponseDto[];
  selectedCategories: CategoryResponseDto[];
  board: GameCategoryBoardResponseDto[];
  currentTurnTeamIndex: number;
  finishedAt?: string;
  createdAt?: string;
  updatedAt?: string;
}